fixed punchline container
There was a bug where bottom right container was not really 50%, and top left container was flowing below bottom right, when user was compressing the browser width. This is now fixed.  Main was changed from grid to flex, and Bottom right container was self-aligned to flex end. Magic! Also punchline container is now  filling 100% of bottom right container. Peeeeerfect!